backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
comm==0.2.2
cryptography==46.0.6
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
executing==2.2.0
faiss-cpu==1.10.0
filelock==3.18.0
Flask==3.1.3
flask-cors==6.0.2
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.49.1
google-genai==1.68.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
orjson==3.10.16
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pinecone==8.1.0
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
postgrest==2.28.3
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.8.1
pydantic_core==2.41.5
Pygments==2.19.1
pyiceberg==0.11.1
PyJWT==2.12.1
PyMuPDF==1.25.4
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.3.0
realtime==2.28.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.7
yarl==1.23.0
zstandard==0.23.0
